@@ -5,9 +5,9 @@
   "type": "library",
   "require": {
     "php": ">=8.2",
-    "symfony/console": "^6.0.0",
-    "symfony/filesystem": "^6.0",
-    "symfony/process": "^6.0"
+    "symfony/console": "^7.0.0",
+    "symfony/filesystem": "^7.0",
+    "symfony/process": "^7.0"
   },
   "suggest": {
     "box-project/box": "Uses box to build phar. Recommended to use box.phar instead of composer."
@@ -27,6 +27,7 @@
     "cotor": {
       "friendsofphp/php-cs-fixer": "^3.1",
       "phpstan/phpstan": "^1.5",
+      "rector/rector": "^1.2",
       "vimeo/psalm": "^4.10"
     }
   }
